# public/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Letta Agent Service
  version: 0.2.0
  description: Local prototype contract using an HttpOnly session cookie. Replace demo credentials with bank SSO/OIDC in production.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: bank_agent_session
  schemas:
    User:
      type: object
      required: [id, username, display_name]
      properties:
        id: { type: string }
        username: { type: string }
        display_name: { type: string }
    Conversation:
      type: object
      required: [id, user_id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string }
        letta_conversation_id: { type: [string, "null"] }
        title: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MessageRequest:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1, maxLength: 32000 }
        request_id: { type: string, description: Stable idempotency key generated by the caller. }
    MessageResponse:
      type: object
      required: [request_id, conversation_id, answer, idempotent_replay]
      properties:
        request_id: { type: string }
        conversation_id: { type: string, format: uuid }
        answer: { type: string }
        duration_ms: { type: [integer, "null"] }
        idempotent_replay: { type: boolean }
security:
  - SessionCookie: []
paths:
  /health:
    get:
      security: []
      summary: Liveness probe
      responses:
        "200": { description: Healthy }
  /v1/auth/login:
    post:
      security: []
      summary: Start a local demo session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200": { description: Authenticated; sets an HttpOnly session cookie }
        "401": { description: Invalid credentials }
        "429": { description: Too many attempts }
  /v1/auth/me:
    get:
      summary: Read the current authenticated identity
      responses:
        "200": { description: Current user }
        "401": { description: Authentication required }
  /v1/auth/logout:
    post:
      summary: Revoke the current session
      responses:
        "204": { description: Logged out }
  /v1/conversations:
    get:
      summary: List the authenticated user's conversations
      responses:
        "200": { description: Conversation list }
    post:
      summary: Create an isolated conversation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 200 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
  /v1/conversations/{conversationId}/messages:
    parameters:
      - name: conversationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Read the authenticated user's audited turn history
      responses:
        "200": { description: Message history }
        "404": { description: Not owned by this user }
    post:
      summary: Run one serialized Agent turn
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageRequest" }
      responses:
        "200":
          description: Completed Agent turn
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "404": { description: Not owned by this user }
        "503": { description: Queue full or dependency unavailable }
  /v1/memory:
    get:
      summary: Inspect only the authenticated user's local MemFS markdown files
      responses:
        "200": { description: User-owned memory files }